requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
amqp==2.3.2               # via kombu
apipkg==1.5               # via execnet
atomicwrites==1.2.1       # via pytest
attrs==18.2.0             # via pytest
aws-requests-auth==0.4.2
backcall==0.1.0           # via ipython
billiard==3.5.0.4         # via celery
boto3==1.9.20
botocore==1.12.21         # via boto3, s3transfer
celery==4.2.1
certifi==2018.8.24        # via requests
chardet==3.0.4
click==7.0                # via pip-tools, towncrier
coverage==4.5.1           # via pytest-cov
cssselect==1.0.3
decorator==4.3.0          # via ipython, traitlets
django-admin-ip-restrictor==0.2.1
django-debug-toolbar==1.10.1
django-environ==0.4.5
django-extensions==2.1.3
django-filter==2.0.0
django-ipware==2.1.0      # via django-admin-ip-restrictor
django-js-asset==1.1.0    # via django-mptt
django-model-utils==3.1.2
django-mptt==0.9.1
django-oauth-toolkit==1.2.0
django-pglocks==1.0.2
django-redis==4.9.0
django-reversion==3.0.0
django==2.1.2
djangorestframework==3.8.2
docopt==0.6.2             # via docopt, notifications-python-client
docutils==0.14            # via botocore, docutils
elasticsearch-dsl==5.4.0
elasticsearch==5.5.3
execnet==1.5.0            # via pytest-xdist
factory-boy==2.11.1
faker==0.9.1              # via factory-boy
flake8-blind-except==0.1.1
flake8-commas==2.0.0
flake8-debugger==3.1.0
flake8-docstrings==1.3.0
flake8-import-order==0.18
flake8-polyfill==1.0.2    # via flake8-docstrings, pep8-naming
flake8-print==3.1.0
flake8-quotes==1.0.0
flake8-string-format==0.2.3
flake8==3.5.0
freezegun==0.3.10
future==0.16.0            # via future, notifications-python-client
gevent==1.3.6
greenlet==0.4.15          # via gevent
gunicorn==19.9.0
idna==2.7                 # via requests
incremental==17.5.0       # via towncrier
ipdb==0.11
ipython-genutils==0.2.0   # via ipython-genutils, traitlets
ipython==7.0.1
jedi==0.13.1              # via ipython
jinja2==2.10              # via towncrier
jmespath==0.9.3           # via boto3, botocore, jmespath
kombu==4.2.1              # via celery
lxml==4.2.5
markupsafe==1.0           # via jinja2
mccabe==0.6.1             # via flake8, mccabe
mohawk==0.3.4
monotonic==1.5            # via notifications-python-client
more-itertools==4.3.0     # via pytest
notifications-python-client==5.2.0
oauthlib==2.1.0           # via django-oauth-toolkit
parso==0.3.1              # via jedi
pep8-naming==0.7.0
pexpect==4.6.0            # via ipython
pickleshare==0.7.5        # via ipython
pip-tools==3.1.0
piprot==0.9.10
pluggy==0.7.1             # via pytest
prompt-toolkit==2.0.5     # via ipython
psycogreen==1.0
psycopg2==2.7.5
ptyprocess==0.6.0         # via pexpect
py==1.6.0                 # via pytest
pycodestyle==2.3.1        # via flake8, flake8-debugger, flake8-import-order, flake8-print, pycodestyle
pydocstyle==2.1.1
pyflakes==1.6.0           # via flake8
pygments==2.2.0           # via ipython, pygments
pyjwt==1.6.4              # via notifications-python-client
pytest-cov==2.6.0
pytest-django==3.4.3
pytest-forked==0.2        # via pytest-xdist
pytest-sugar==0.9.1
pytest-xdist==1.23.2
pytest==3.8.2
python-dateutil==2.7.3
pytz==2018.5              # via celery, django
pyyaml==3.13
raven==6.9.0
redis==2.10.6             # via django-redis
requests-futures==0.9.7   # via piprot
requests-mock==1.5.2
requests==2.19.1
requests_toolbelt==0.8.0
s3transfer==0.1.13        # via boto3
simplegeneric==0.8.1      # via ipython, simplegeneric
six==1.11.0               # via django-extensions, elasticsearch-dsl, faker, flake8-print, freezegun, mohawk, more-itertools, pip-tools, piprot, prompt-toolkit, pydocstyle, pytest, pytest-xdist, python-dateutil, requests-mock, six, traitlets
snowballstemmer==1.2.1    # via pydocstyle, snowballstemmer
sqlparse==0.2.4           # via django-debug-toolbar, sqlparse
termcolor==1.1.0          # via pytest-sugar, termcolor
text-unidecode==1.2       # via faker
toml==0.10.0              # via towncrier
towncrier==18.6.0
traitlets==4.3.2          # via ipython, traitlets
urllib3==1.23             # via botocore, elasticsearch, requests
vine==1.1.4               # via amqp
wcwidth==0.1.7            # via prompt-toolkit, wcwidth
werkzeug==0.14.1
whitenoise==4.1
